+++ tomcat/src/main/java/org/apache/catalina/Session.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.catalina;
 
 import java.util.HashMap;
 import java.util.Map;
